// server/data/interfaces/prisonApi/FullStatus.ts
export default interface FullStatus {
  nomsId: string
  establishmentCode: string
  bookingId: number
  givenName1: string
  lastName: string
  dateOfBirth: string
  gender: string
  englishSpeaking: boolean
  bookingBeginDate: string
  releaseDate: string
  categoryCode: string
  communityStatus: string
  legalStatus: string
  establishmentName: string
}