// server/data/interfaces/prisonApi/HistoryForLocationItem.ts
export default interface HistoryForLocationItem {
  bookingId: number
  livingUnitId: number
  assignmentDate: string
  assignmentDateTime: string
  assignmentReason: string
  assignmentEndDate: string
  assignmentEndDateTime: string
  agencyId: string
  description: string
  bedAssignmentHistorySequence: number
  movementMadeBy: string
  offenderNo: string
}